requirements.txt
# ============================================
# RAG个性化智能学习伴侣系统 - 依赖配置
# Python 3.12.6
# ============================================

# === 核心框架 ===
streamlit>=1.40.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
llama-index>=0.11.0
llama-parse>=0.5.0

# === 向量数据库 ===
chromadb>=0.5.0

# === Embedding & Reranker ===
google-generativeai>=0.8.0

# === 文档解析 ===
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0
pdfplumber>=0.11.0

# === 数据处理 ===
pandas>=2.0.0
numpy>=1.26.0

# === BM25检索 ===
rank-bm25>=0.2.2

# === 知识图谱可视化 (Phase 2) ===
streamlit-agraph>=0.0.45

# === 可视化组件 ===
plotly>=5.24.0

# === 工具库 ===
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
tenacity>=8.2.0

# === 密码哈希 ===
bcrypt>=4.1.0

# === 登录状态持久化 ===
extra-streamlit-components>=0.1.60

# === 开发工具 ===
pytest>=8.0.0
black>=24.0.0
